cleaned up things after merge
