--- config/mimic3.ini
@@ -1,21 +1,0 @@
-[constants]
-freq_band=[0.5, 8.0]
-sim=0.6
-snr=2.0
-hr_freq_band=[0.667, 3.0]
-hr_delta=1/6
-abp_min_bounds=[20, 130]
-abp_max_bounds=[50, 225]
-windowsize=1
-ma_perc=20
-beat_sim=0.2
-has_abp=True
-
-[checks]  # only requires ppg
-snr
-hr
-flat
-beat
-bp
-sim
-hr_delta
