--- config/prod.ini
@@ -1,22 +1,0 @@
-[constants]
-freq_band=[0.5, 8.0]
-sim=0.6
-snr=2.0
-hr_freq_band=[0.667, 3.0]
-hr_delta=1/6
-abp_min_bounds=[20, 130]
-abp_max_bounds=[50, 225]
-windowsize=1
-ma_perc=20
-beat_sim=0.2
-
-[ppg checks]  # only requires ppg
-nan
-snr
-hr
-flat
-beat
-
-[abp checks]  # only requires abp
-
-[checks]  # requires ppg and abp
