+++ scripts/mimic3_data_locator.py
@@ -1,5 +1,5 @@
 from database_tools.tools import MimicDataLocator
-from database_tools.preprocessing.utils import build_data_directory
+from database_tools.filtering.utils import build_data_directory
 
 repo_dir = '/home/cam/Documents/database_tools/'
 data_dir = build_data_directory(repo_dir + 'data/', 'mimic3')
